Remove unnecessary UNDEFINED entry type from PerformanceObserver (#43701)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43701

Changelog: [internal]

We were defining an `UNDEFINED` type of entry that should never happen in practice and we were using unnecessarily to signal "no type" where an optional type would be more suitable. Most importantly, **we were incorrectly allocating a buffer for entries of this type**.

This removes that type and the unnecessary buffer.

Reviewed By: rshest

Differential Revision: D55478890

fbshipit-source-id: 145210a9c4e2614a342f2d913b9eb6b3d62f676f