Do not attempt to call `Class.forName` in OSS (#43816)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43816

We should not attempt to load generated classes by the Annotation processor in OSS because we simply don't run it, so those classes will fail to load. I'm short-circuiting the logic here.

This is sustainability work as we got a report for this in OSS a while ago and I never got the time to work on it.

Changelog:
[Internal] [Changed] - Do not attempt to call Class.forName in OSS

Reviewed By: rshest

Differential Revision: D55693479

fbshipit-source-id: 3ec84e2c7940011b48f354058b5099b46065166d