Use different device IDs for modern and legacy backend (#43968)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/43968

Changelog: [Internal]

When going from a Fusebox build to a non-Fusebox build of the same app, users can accidentally connect a previously opened Fusebox frontend to a non-Fusebox backend (or vice versa).

To prevent this, here we assign a distinct "device ID" to the Fusebox backend on both Android and iOS. This will prevent a mismatched CDT frontend from reconnecting to the app after a backend change, forcing the user to close and reopen the debugger.

Reviewed By: hoxyq

Differential Revision: D55870800

fbshipit-source-id: 8552e009ef1e43b512d35035c98d1a859e3ccf91